+++ src/EventStore.Core.Tests/Services/PersistentSubscription/PersistentSubscriptionMessageParkerTests.cs
@@ -28,7 +28,7 @@ namespace EventStore.Core.Tests.Services.PersistentSubscription {
 		}
 
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		public class given_parked_stream_does_not_exist<TLogFormat, TStreamId> : TestFixtureWithExistingEvents<TLogFormat, TStreamId> {
 			private PersistentSubscriptionMessageParker _messageParker;
 			private TaskCompletionSource<bool> _done = new TaskCompletionSource<bool>();
@@ -50,7 +50,7 @@ namespace EventStore.Core.Tests.Services.PersistentSubscription {
 		}
 
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		public class given_parked_messages_and_no_truncate_before<TLogFormat, TStreamId> : TestFixtureWithExistingEvents<TLogFormat, TStreamId> {
 			private PersistentSubscriptionMessageParker _messageParker;
 			private string _streamId = Guid.NewGuid().ToString();
@@ -77,7 +77,7 @@ namespace EventStore.Core.Tests.Services.PersistentSubscription {
 
 
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		public class given_parked_messages_and_half_are_truncated<TLogFormat, TStreamId> : TestFixtureWithExistingEvents<TLogFormat, TStreamId> {
 			private TaskCompletionSource<bool> _done = new TaskCompletionSource<bool>();
 			private PersistentSubscriptionMessageParker _messageParker;
@@ -104,7 +104,7 @@ namespace EventStore.Core.Tests.Services.PersistentSubscription {
 		}
 
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		public class given_parked_messages_and_all_are_truncated<TLogFormat, TStreamId> : TestFixtureWithExistingEvents<TLogFormat, TStreamId> {
 			private PersistentSubscriptionMessageParker _messageParker;
 			private string _streamId = Guid.NewGuid().ToString();
@@ -137,7 +137,7 @@ namespace EventStore.Core.Tests.Services.PersistentSubscription {
 		}
 
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		public class given_a_message_is_parked<TLogFormat, TStreamId> : TestFixtureWithExistingEvents<TLogFormat, TStreamId> {
 			private PersistentSubscriptionMessageParker _messageParker;
 			private string _streamId = Guid.NewGuid().ToString();
@@ -161,7 +161,7 @@ namespace EventStore.Core.Tests.Services.PersistentSubscription {
 		}
 
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		public class given_messages_are_parked_and_then_replayed<TLogFormat, TStreamId> : TestFixtureWithExistingEvents<TLogFormat, TStreamId> {
 			private PersistentSubscriptionMessageParker _messageParker;
 			private string _streamId = Guid.NewGuid().ToString();
@@ -194,7 +194,7 @@ namespace EventStore.Core.Tests.Services.PersistentSubscription {
 		}
 
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		public class given_message_parked_after_parked_messages_are_replayed<TLogFormat, TStreamId> : TestFixtureWithExistingEvents<TLogFormat, TStreamId> {
 			private PersistentSubscriptionMessageParker _messageParker;
 			private string _streamId = Guid.NewGuid().ToString();
@@ -229,7 +229,7 @@ namespace EventStore.Core.Tests.Services.PersistentSubscription {
 		}
 
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		public class given_read_backwards_fails_when_getting_stats<TLogFormat, TStreamId> : TestFixtureWithExistingEvents<TLogFormat, TStreamId> {
 			private PersistentSubscriptionMessageParker _messageParker;
 			private string _streamId = Guid.NewGuid().ToString();
@@ -267,7 +267,7 @@ namespace EventStore.Core.Tests.Services.PersistentSubscription {
 		}
 
 		[TestFixture(typeof(LogFormat.V2), typeof(string))]
-		[TestFixture(typeof(LogFormat.V3), typeof(long))]
+		[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 		public class given_read_forwards_fails_when_getting_stats<TLogFormat, TStreamId> : TestFixtureWithExistingEvents<TLogFormat, TStreamId> {
 			private PersistentSubscriptionMessageParker _messageParker;
 			private string _streamId = Guid.NewGuid().ToString();
