+++ src/EventStore.Core.Tests/Services/Storage/Scavenge/when_deleting_single_stream_spanning_through_2_chunks_in_db_with_2_chunks.cs
@@ -4,7 +4,7 @@ using NUnit.Framework;
 
 namespace EventStore.Core.Tests.Services.Storage.Scavenge {
 	[TestFixture(typeof(LogFormat.V2), typeof(string))]
-	[TestFixture(typeof(LogFormat.V3), typeof(long))]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint))]
 	public class when_deleting_single_stream_spanning_through_2_chunks_in_db_with_2_chunks<TLogFormat, TStreamId> : ReadIndexTestScenario<TLogFormat, TStreamId> {
 		private EventRecord _event3;
 		private EventRecord _event4;
